Corrected TTL handling

+++ src/integrations/dns.rs
@@ -18,7 +18,7 @@ pub struct DnsConfiguration {
     /// FQDN, used for resolving the endpoints A/AAAA entries for stickyness
     pub record: String,
     /// if this endpoint is selected, the TTL will be applied to the entries part of the backend record
-    pub ttl: u64,
+    pub ttl: u16,
     /// the DNS record will be resolved by this resolver
     pub resolver: String,
 }
